CSELEC-3-main/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pymongo==4.6.2
python-dotenv==1.0.1
gunicorn==21.2.0
flask-caching==2.1.0 